+++ extensions/iceworks-app/src/quickPicks/options.ts
@@ -1,6 +1,5 @@
-import * as vscode from 'vscode';
-import { getProjectType, checkIsPegasusProject, checkIsNotTarget } from '@iceworks/project-service';
-import { checkIsAliInternal, checkIsO2 } from '@iceworks/common-service';
+import { getProjectType, checkIsPegasusProject, checkIsTargetProjectType } from '@iceworks/project-service';
+import { checkIsAliInternal, checkIsO2, checkIsInstalledDoctor } from '@iceworks/common-service';
 import i18n from '../i18n';
 
 export default [
@@ -13,21 +12,23 @@ export default [
       return !isO2;
     },
   },
-  // {
-  //   label: i18n.format('extension.iceworksApp.showEntriesQuickPick.customScaffold.label'),
-  //   detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.customScaffold.detail'),
-  //   command: 'iceworks-project-creator.custom-scaffold.start',
-  // },
   {
-    label: i18n.format('extension.iceworksApp.showEntriesQuickPick.openDashboard.label'),
-    detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.openDashboard.detail'),
+    label: i18n.format('extension.iceworksApp.showEntriesQuickPick.dashboard.label'),
+    detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.dashboard.detail'),
+    command: 'iceworksApp.dashboard.start',
+    async condition() {
+      const isTargetProjectType = await checkIsTargetProjectType();
+      return isTargetProjectType;
+    },
+  },
+  {
+    label: i18n.format('extension.iceworksApp.showEntriesQuickPick.doctor.label'),
+    detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.doctor.detail'),
     command: 'iceworks-doctor.dashboard',
     async condition() {
-      const doctorExtension = vscode.extensions.getExtension('iceworks-team.iceworks-doctor');
-      const isTargetProject = !(await checkIsNotTarget());
-      // TODO disable Doctor in O2: too large causes GC to be packaged
-      const isO2 = checkIsO2();
-      return !isO2 && doctorExtension && isTargetProject;
+      const isInstalledDoctor = checkIsInstalledDoctor();
+      const isTargetProjectType = await checkIsTargetProjectType();
+      return isInstalledDoctor && isTargetProjectType;
     },
   },
   {
@@ -35,7 +36,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.runDebug.detail'),
     command: 'iceworksApp.editorMenu.runDebug',
     async condition() {
-      return !(await checkIsNotTarget());
+      return await checkIsTargetProjectType();
     },
   },
   {
@@ -43,7 +44,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.DefPublish.detail'),
     command: 'iceworksApp.editorMenu.DefPublish',
     async condition() {
-      return (await checkIsAliInternal()) && !(await checkIsNotTarget());
+      return (await checkIsAliInternal()) && await checkIsTargetProjectType();
     },
   },
   {
@@ -51,7 +52,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.generatePage.detail'),
     command: 'iceworks-material-helper.page-generator.start',
     async condition() {
-      return !(await checkIsNotTarget()) && !(await checkIsPegasusProject());
+      return await checkIsTargetProjectType() && !(await checkIsPegasusProject());
     },
   },
   {
@@ -59,7 +60,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.createPage.detail'),
     command: 'iceworks-material-helper.page-creator.start',
     async condition() {
-      return !(await checkIsNotTarget()) && !(await checkIsPegasusProject());
+      return await checkIsTargetProjectType() && !(await checkIsPegasusProject());
     },
   },
   {
@@ -67,7 +68,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.materialImport.detail'),
     command: 'iceworks-material-helper.material-importer.start',
     async condition() {
-      return !(await checkIsNotTarget());
+      return await checkIsTargetProjectType();
     },
   },
   {
@@ -86,7 +87,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.createComponent.detail'),
     command: 'iceworks-material-helper.component-creator.start',
     async condition() {
-      return !(await checkIsNotTarget()) && !(await checkIsPegasusProject());
+      return await checkIsTargetProjectType() && !(await checkIsPegasusProject());
     },
   },
   {
@@ -94,7 +95,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.showMaterialDocs.detail'),
     command: 'iceworks-material-helper.showMaterialDocs',
     async condition() {
-      return !(await checkIsNotTarget());
+      return await checkIsTargetProjectType();
     },
   },
   {
@@ -102,7 +103,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.reinstall.detail'),
     command: 'iceworksApp.nodeDependencies.reinstall',
     async condition() {
-      return !(await checkIsNotTarget());
+      return await checkIsTargetProjectType();
     },
   },
   {
@@ -110,7 +111,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.addDepsAndDevDeps.detail'),
     command: 'iceworksApp.nodeDependencies.addDepsAndDevDeps',
     async condition() {
-      return !(await checkIsNotTarget());
+      return await checkIsTargetProjectType();
     },
   },
   {
@@ -118,7 +119,7 @@ export default [
     detail: i18n.format('extension.iceworksApp.showEntriesQuickPick.runBuild.detail'),
     command: 'iceworksApp.editorMenu.runBuild',
     async condition() {
-      return !(await checkIsNotTarget());
+      return await checkIsTargetProjectType();
     },
   },
   {