+++ extensions/iceworks-app/web/src/pages/Dashboard/models/componentDependencies.ts
@@ -1,0 +1,18 @@
+import callService from '@/callService';
+
+export default {
+  state: {
+    data: [],
+    inited: false,
+  },
+  reducers: {
+  },
+  effects: () => ({
+    async refresh() {
+      this.setState({ inited: true });
+      this.setState({
+        data: await callService('project', 'getComponentDependencies'),
+      });
+    },
+  }),
+};