+++ extensions/iceworks-doctor/src/getScanReport.ts
@@ -2,6 +2,7 @@ import { Doctor } from '@iceworks/doctor';
 import { projectPath, getProjectType, getProjectLanguageType } from '@iceworks/project-service';
 import getRecorder from './getRecorder';
 import setDiagnostics from './setDiagnostics';
+import storage from './storage';
 
 const doctor = new Doctor({ ignore: ['.vscode', '.ice', 'mocks', '.eslintrc.js', 'webpack.config.js'] });
 
@@ -23,6 +24,9 @@ export default async (options) => {
     });
 
     report = await doctor.scan(targetPath, scanOption);
+
+    // store the latest result locally
+    storage.saveReport(report);
 
     // Set VS Code problems
     setDiagnostics(report.securityPractices, true);