+++ extensions/iceworks-doctor/src/storage.ts
@@ -1,0 +1,3 @@
+import { DoctorStorage } from '@iceworks/storage';
+
+export default new DoctorStorage();