+++ packages/common-service/src/index.ts
@@ -32,7 +32,6 @@ export const CONFIGURATION_KEY_NPM_REGISTRY = 'npmRegistry';
 export const CONFIGURATION_KEY_MATERIAL_SOURCES = 'materialSources';
 export const CONFIGURATION_KEY_GENERATE_PAGE_PATH = 'generatePagePath';
 export const CONFIGURATION_KEY_GENERATE_COMPONENT_PATH = 'generateComponentPath';
-
 export const CONFIGURATION_SECTION_PCKAGE_MANAGER = `${CONFIGURATION_SECTION}.${CONFIGURATION_KEY_PCKAGE_MANAGER}`;
 export const CONFIGURATION_SECTION_NPM_REGISTRY = `${CONFIGURATION_SECTION}.${CONFIGURATION_KEY_NPM_REGISTRY}`;
 export const CONFIGURATION_SETION_MATERIAL_SOURCES = `${CONFIGURATION_SECTION}.${CONFIGURATION_KEY_MATERIAL_SOURCES}`;
@@ -85,6 +84,10 @@ export async function checkPathExists(p: string, folderName?: string): Promise<b
     p = path.join(p, folderName);
   }
   return await fse.pathExists(p);
+}
+
+export function openInExternalFinder(url) {
+  vscode.env.openExternal(vscode.Uri.file(url));
 }
 
 export function saveDataToSettingJson(section: string, data: any, configurationTarget: boolean = true): void {
@@ -246,6 +249,20 @@ async function autoSetNpmRegistryConfiguration(globalState: vscode.Memento, isAl
   });
 }
 
+export function isYarnPackageManager(): boolean {
+  const packageManager = getDataFromSettingJson('packageManager', 'npm');
+  const isYarn = packageManager === 'yarn';
+  return isYarn;
+}
+
+export function getAddDependencyAction(): 'add' | 'install' {
+  return isYarnPackageManager() ? 'add' : 'install';
+}
+
+export function getUpdateDependencyAction(): 'upgrade' | 'update' {
+  return isYarnPackageManager() ? 'upgrade' : 'update';
+}
+
 export function createNpmCommand(action: string, target: string = '', extra: string = ''): string {
   const packageManager = getDataFromSettingJson('packageManager', 'npm');
   let registry = '';
@@ -369,7 +386,7 @@ export const getFolderLanguageType = (templateSourceSrcPath) => {
   });
 
   return index >= 0 ? 'ts' : 'js';
-};
+}
 
 /**
  * Install materials dependencies
@@ -400,14 +417,15 @@ export const bulkInstallMaterialsDependencies = async function (
       return `${packageName}@${version}`;
     });
 
+    const addDependencyAction = getAddDependencyAction(); // `add` or `install`
     const terminal = getIceworksTerminal();
     terminal.show();
     terminal.sendText(`cd '${projectPath}'`, true);
-    terminal.sendText(createNpmCommand('install', deps.join(' '), '--save'), true);
+    terminal.sendText(createNpmCommand(addDependencyAction, deps.join(' '), '--save'), true);
   } else {
     return [];
   }
-};
+}
 
 export const bulkDownloadMaterials = async function (
   materials: IMaterialPage[] | IMaterialBlock[],
@@ -463,7 +481,8 @@ export const bulkDownloadMaterials = async function (
       }
     })
   );
-};
+}
+
 export function openMaterialsSettings() {
   if (vscode.extensions.getExtension('iceworks-team.iceworks-app')) {
     executeCommand('iceworksApp.configHelper.start', 'iceworks.materialSources');
@@ -493,4 +512,9 @@ export function getFolderExistsTime(folderPath: string) {
   const curTime = new Date();
   const existsTime = ((curTime.getTime() - birthtime.getTime()) / (60 * 1000))
   return existsTime;
+}
+
+export function checkIsInstalledDoctor() {
+  const doctorExtension = vscode.extensions.getExtension('iceworks-team.iceworks-doctor');
+  return !!doctorExtension;
 }