+++ packages/project-service/src/index.ts
@@ -2,49 +2,50 @@ import * as vscode from 'vscode';
 import * as fsExtra from 'fs-extra';
 import { downloadAndGenerateProject } from '@iceworks/generate-project';
 import { checkPathExists, getDataFromSettingJson, CONFIGURATION_KEY_NPM_REGISTRY } from '@iceworks/common-service';
-import { checkIsTargetProject, getProjectType as originGetProjectType, getProjectFramework as originGetProjectFramework } from '@iceworks/project-utils';
+import { checkIsTargetProjectType as orginCheckIsTargetProjectType, checkIsTargetProjectFramework as orginCheckIsTargetProjectFramework, getProjectType as originGetProjectType, getProjectFramework as originGetProjectFramework } from '@iceworks/project-utils';
 import * as simpleGit from 'simple-git/promise';
 import * as path from 'path';
-import axios from 'axios';
-import { ALI_EMAIL, ALI_GITLAB_URL } from '@iceworks/constant';
-import {
-  generatorCreatetaskUrl,
-  generatorTaskResultUrl,
-  applyRepositoryUrl,
-  GeneratorTaskStatus,
-  projectPath,
-  jsxFileExtnames,
-} from './constant';
+import { ALI_GITLAB_URL, ALI_DIP_PRO, ALI_DEF_WORK_URL } from '@iceworks/constant';
+import { projectPath, jsxFileExtnames } from './constant';
+import { generatorCreatetask, getGeneratorTaskStatus, applyRepository, getBasicInfo } from './def';
+import { getInfo } from './git';
 import i18n from './i18n';
+import { getProjectPackageJSON } from './utils';
 import { IDEFProjectField, IProjectField } from './types';
 
 export * from './constant';
+export * from './dependency';
 
 export async function autoSetContext() {
   const isPegasus = await checkIsPegasusProject();
   const languageType = await getProjectLanguageType();
   const type = await getProjectType();
   const framework = await getProjectFramework();
-  const isNotTarget = await checkIsNotTarget();
-  vscode.commands.executeCommand('setContext', 'iceworks:projectIsNotTarget', isNotTarget);
+  const isNotTargetType = !await checkIsTargetProjectType();
+  const isNotTargetFramework = !await checkIsTargetProjectFramework();
+  vscode.commands.executeCommand('setContext', 'iceworks:projectIsNotTargetType', isNotTargetType);
+  vscode.commands.executeCommand('setContext', 'iceworks:projectIsNotTargetFramework', isNotTargetFramework);
   vscode.commands.executeCommand('setContext', 'iceworks:projectIsPegasus', isPegasus);
   vscode.commands.executeCommand('setContext', 'iceworks:projectLanguageType', languageType);
   vscode.commands.executeCommand('setContext', 'iceworks:projectType', type);
   vscode.commands.executeCommand('setContext', 'iceworks:projectFramework', framework);
 }
 
-export async function checkIsNotTarget() {
-  let isNotTarget = false;
-  if (!vscode.workspace.rootPath) {
-    isNotTarget = true;
-  } else {
-    try {
-      isNotTarget = !await checkIsTargetProject(projectPath);
-    } catch (e) {
-      isNotTarget = true;
-    }
+export async function checkIsTargetProjectType() {
+  return await orginCheckIsTargetProjectType(projectPath);
+}
+
+export async function checkIsTargetProjectFramework() {
+  return await orginCheckIsTargetProjectFramework(projectPath);
+}
+
+export async function getFeedbackLink() {
+  const framework = await getProjectFramework();
+  if (framework === 'icejs') {
+    return 'https://c.tb.cn/F3.ZpKQYk';
+  } else if (framework === 'rax-app') {
+    return 'https://c.tb.cn/F3.ZLhGNW';
   }
-  return isNotTarget;
 }
 
 export async function getProjectLanguageType() {
@@ -52,8 +53,8 @@ export async function getProjectLanguageType() {
 
   const framework = await getProjectFramework();
   let isTypescript = false;
-  if (framework === 'icejs') {
-    // icejs 都有 tsconfig，因此需要通过 src/app.js[x] 进一步区分
+
+  if (framework !== 'unknown') {
     const hasAppJs =
       fsExtra.existsSync(path.join(projectPath, 'src/app.js')) ||
       fsExtra.existsSync(path.join(projectPath, 'src/app.jsx'));
@@ -67,6 +68,41 @@ export async function getProjectLanguageType() {
 
 export async function getProjectType() {
   return await originGetProjectType(projectPath);
+}
+
+export async function getProjectBaseInfo() {
+  const { name, description } = await getProjectPackageJSON();
+  const type = await getProjectType();
+  const framework = await getProjectFramework();
+  const feedbackLink = await getFeedbackLink();
+  return {
+    name,
+    description,
+    type,
+    framework,
+    path: projectPath,
+    feedbackLink,
+  };
+}
+
+export async function getProjectGitInfo() {
+  const info = await getInfo(projectPath);
+  const repository = info.repository
+    .replace(/^git@/, 'https://')
+    .replace(/\.git/, '')
+    .replace(/\.com:/, '.com/');
+  const [,,, group, project] = repository.split('/');
+  return { ...info, repository, group, project };
+}
+
+export async function getProjectDefInfo(clientToken: string) {
+  const { group, project } = await getProjectGitInfo();
+  const info = await getBasicInfo(`${group}/${project}`, clientToken);
+  return {
+    ...info,
+    defUrl: `${ALI_DEF_WORK_URL}/app/${info.id}`,
+    idpUrl: ALI_DIP_PRO,
+  };
 }
 
 export async function checkIsPegasusProject() {
@@ -84,14 +120,6 @@ export async function checkIsPegasusProject() {
 
 export async function getProjectFramework() {
   return await originGetProjectFramework(projectPath);
-}
-
-export async function getPackageJSON(packagePath: string): Promise<any> {
-  const packagePathIsExist = await fsExtra.pathExists(packagePath);
-  if (!packagePathIsExist) {
-    throw new Error(i18n.format('package.projectService.index.packageNotFound'));
-  }
-  return await fsExtra.readJson(packagePath);
 }
 
 export function getIceVersion(packageJSON): string {
@@ -120,7 +148,11 @@ export async function getFolderPath(openLabel = 'Open'): Promise<string | undefi
 }
 
 export async function createProject(projectField: IProjectField): Promise<string> {
-  const { projectPath: setProjectPath, projectName, scaffold, ejsOptions } = projectField;
+  const { projectPath: setProjectPath, projectName, scaffold } = projectField;
+  let ejsOptions = {};
+  if (projectField.ejsOptions) {
+    ejsOptions = modifyEjsOptions(projectField.ejsOptions);
+  }
   const projectDir: string = path.join(setProjectPath, projectName);
   const isProjectDirExists = await checkPathExists(projectDir);
   if (isProjectDirExists) {
@@ -173,100 +205,31 @@ async function cloneRepositoryToLocal(projectDir, group, project): Promise<void>
   await simpleGit().clone(repoPath, projectDir);
 }
 
-async function generatorCreatetask(field: IDEFProjectField) {
-  const { empId, account, group, project, gitlabToken, scaffold, clientToken, ejsOptions } = field;
-  const projectType = field.source.type;
-  const { description, source } = scaffold;
-  const { npm } = source;
-  let generatorId = 6;
-  if (projectType === 'rax') {
-    generatorId = 5;
-  }
-  const response = await axios.post(generatorCreatetaskUrl, {
-    group,
-    project,
-    description,
-    trunk: 'master',
-    generator_id: generatorId,
-    schema_data: {
-      npmName: npm,
-      ...ejsOptions,
-    },
-    gitlab_info: {
-      id: empId,
-      token: gitlabToken,
-      name: account,
-      email: `${account}@${ALI_EMAIL}`,
-    },
-    emp_id: empId,
-    client_token: clientToken,
-  });
-  console.log('generatorCreatetaskResponse', response);
-  if (response.data.error) {
-    throw new Error(response.data.error);
-  }
-  return response;
-}
+function modifyEjsOptions(ejsOptions) {
+  let enableMPA = false;
+  let enablePHA = false;
+  let targets = [];
 
-function getGeneratorTaskStatus(taskId: number, clientToken: string): Promise<any> {
-  return new Promise((resolve, reject) => {
-    const interval = setInterval(async () => {
-      try {
-        const response = await axios.get(`${generatorTaskResultUrl}/${taskId}`, {
-          params: {
-            need_generator: true,
-            client_token: clientToken,
-          },
-        });
-        console.log('generatorTaskResultResponse', response);
-        const {
-          data: { status },
-          error,
-        } = response.data;
-        if (error) {
-          reject(new Error(error));
-        }
-        if (status !== GeneratorTaskStatus.running && status !== GeneratorTaskStatus.Created) {
-          clearInterval(interval);
-          if (status === GeneratorTaskStatus.Failed) {
-            reject(new Error(i18n.format('package.projectService.index.DEFOutTime', { taskId })));
-          }
-          if (status === GeneratorTaskStatus.Timeout) {
-            reject(new Error(i18n.format('package.projectService.index.DEFOutTime', { taskId })));
-          }
-          if (status === GeneratorTaskStatus.Success) {
-            resolve();
-          }
-        }
-      } catch (error) {
-        reject(error);
-      }
-    }, 1000);
-  });
-}
+  const { appType } = ejsOptions;
 
-async function applyRepository(field: IDEFProjectField) {
-  const { empId, group, project, scaffold, clientToken, source } = field;
-  const { description } = scaffold;
-  const reason = '';
-  const user = [];
-  let pubtype = 1; // default publish type: assets
-  if (source.type === 'rax') {
-    pubtype = 6;
+  if (appType === 'web-mpa') {
+    enableMPA = true;
+    enablePHA = true;
+    targets = ['web'];
+  } else if (appType === 'miniapp') {
+    targets = ['web', 'miniapp', 'wechat-miniprogram'];
+  } else if (appType === 'kraken-mpa') {
+    enableMPA = true;
+    targets = ['web', 'kraken'];
+  } else if (appType === 'weex-mpa') {
+    enableMPA = true;
+    targets = ['web', 'weex'];
   }
-  const response = await axios.post(applyRepositoryUrl, {
-    emp_id: empId,
-    group,
-    project,
-    description,
-    reason,
-    pubtype,
-    user,
-    client_token: clientToken,
-  });
-  console.log('applyRepositoryResponse', response);
-  if (response.data.error) {
-    throw new Error(response.data.error);
-  }
-  return response;
+
+  return {
+    ...ejsOptions,
+    mpa: enableMPA,
+    pha: enablePHA,
+    targets,
+  };
 }