+++ extensions/iceworks-app/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.7.4
+
+- refactor: page&component tree display logic
+- fix: rax app language type is always ts
+
 ## 0.7.3
 
 - feat: add menus to explorer