+++ extensions/iceworks-doctor/CHANGELOG.md
@@ -1,11 +1,24 @@
 # Change Log
 
+## 0.3.1
+
+- refactir: remove project info
+
+## 0.3.0
+
+- refactor: remove project info
+- fix: rax-app language type is always ts
+
 ## 0.2.2
+
 - feat: support vue project.
+
 ## 0.2.1
+
 - feat: update [@iceworks/doctor](https://www.npmjs.com/package/@iceworks/doctor).
 
 ## 0.2.0
+
 - feat: update [@iceworks/doctor](https://www.npmjs.com/package/@iceworks/doctor) and scan reports.
 
 ## 0.1.7