+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.4.7
+
+- docs: delete Configure Scaffold section
+- fix: rax-app language type is always ts
+
 ## 0.4.6
 
 - feat: record DAU for O2